# dqops/src/main/resources/api-dqo-cloud-swagger.yaml
...
swagger: "2.0"
info:
  description: "DQOps Cloud API"
  version: "v1"
  title: "DQOps Cloud"
  termsOfService: "https://dqops.com/terms-of-service"
  contact:
    name: "DQOps Cloud Support"
    url: "https://dqops.com/"
    email: "support@dqops.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "api.dqops.com"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /api/v1/apikey/issueapikey:
    post:
      tags:
      - "ApiKeyRequest"
      summary: "issueApiKey"
      description: "Issues an API Key by the DQO Cloud UI, this operation can also\
        \ generate a new fresh API key."
      operationId: "issueApiKey"
      consumes:
      - "text/plain"
      produces:
      - "text/plain"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
      security:
      - api_key: []
      - firebase_auth: []
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "APIKey"
    in: "header"
  basic:
    type: "basic"
  firebase_auth:
    type: "oauth2"
    authorizationUrl: ""
    flow: "implicit"
...